Collapse the four wire formats behind one decodeSession (#505)
The decode path was a single `async` function that fetched for itself: two
lines of I/O and ninety of format logic, and because the whole thing was
async and did its own fetching, **none of that format logic was reachable
from a test**. That defect, more than the duplication, is what candidate 5
exists to fix.

**One interface, one adapter per format.** `WIRE_FORMATS` in
`js/sessionCodec.js` holds the four -- `session`, `juiceboxURL`, `juicebox`,
`query` -- each a pair of "is this mine?" and "decode it into the shared
context", folded in order. The order is not decoration: `juiceboxURL`
rewrites the query the later adapters read, `juicebox` overwrites what
`session` produced, and `query` overwrites both when it names a map. That is
the precedence the old statement order expressed, and the golden file pins
it. A fifth format is an entry in that array and nothing else.

**The decoder does no I/O.** The session-URL fetch and the bit.ly expansion
arrive as injected loaders, so `test/testSessionDecode.js` drives every
format -- including the session-URL path, and the case that motivated the
injection, a URL whose *contents* need sniffing -- from string literals and a
fake loader. The suite's default loaders throw, so a format that starts
reaching for the wire fails loudly.

`extractConfig` stays internal and unexported: it is now the injection shim,
and it keeps `expandURL` company as the I/O edge. Adding a decoder to the
public surface would create a second, weaker contract we would owe
compatibility to (ADR-0006 decision 9).

Shortcut expansion stays duplicated, as the ticket asks -- collapsing it is
the normalize stage, candidate 9. The primitives move here only because
`decodeQuery` expands at three call sites and leaving them behind would make
the two modules import each other.

Everything moved is moved verbatim, `var`s and semicolons and all: the golden
file is the only thing standing between this refactor and a silent change to
a format users have pasted into papers, and a verbatim move is the one kind
it can vouch for.

`test/__snapshots__/testDecoderGolden.js.snap` is unmodified. 684 tests pass.
